+++ NEWS.md
@@ -1,3 +1,15 @@
+# pushover 1.1.0.9000 (development version)
+
+## New Features
+
+* `pushover()`
+  * `glue()` is now evaluated prior to character limit checks
+  * message priority can now be specified as a string (e.g., "normal") rather than integer
+* `update_glance`
+  * `glue()` formatting for `title`, `text`, and `subtext` values
+  * Non-integer `count` and `percent` values are now coerced to integer rather than aborting
+
+
 # pushoverr 1.1.0 (2021-11-13)
 
 ## New Features